// src/app/back-service/model/jobDTO.ts
/**
 * Member Management Service API
 * MMS is application intended to register and manager members, record attendance and track contribution of each members. It also allow to schedule meetings and send notification to members of the church.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: metyh2@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface JobDTO { 
    jobId :string;
    fromDate :string;
    toDate :string;
    name :string;
    file :string;
    dateTime? :string;
    total_Rows :number;
    processed_Rows :number;
    skipped_Rows :number;
    status :string;
}